+++ src/actions/ActionTypes.js
@@ -10,7 +10,6 @@
 export const COLUMN_REORDER_START = 'COLUMN_REORDER_START';
 export const COLUMN_REORDER_END = 'COLUMN_REORDER_END';
 export const COLUMN_REORDER_MOVE = 'COLUMN_REORDER_MOVE';
-export const COLUMN_RESIZE = 'COLUMN_RESIZE';
 export const INITIALIZE = 'INITIALIZE';
 export const PROP_CHANGE = 'PROP_CHANGE';
 export const SCROLL_START = 'SCROLL_START';
